+++ example/lib/main.dart
@@ -9,7 +9,6 @@ import 'screens/collection_tab.dart';
 import 'screens/home_tab.dart';
 import 'screens/shop_tab.dart';
 import 'screens/search_tab.dart';
-// import 'screens/checkout_page.dart';
 import 'screens/orders_tab.dart';
 
 Future<void> main() async {
@@ -57,7 +56,6 @@ class MyHomePageState extends State<MyHomePage> {
     const SearchTab(),
     const ShopTab(),
     const BlogTab(),
-    // const CheckoutPage(),
     const CartTab(),
     const OrdersTab(),
     const AuthTab(),