+++ lib/graphql_operations/storefront/mutations/customer_access_token_create.dart
@@ -6,6 +6,11 @@ mutation customerAccessTokenCreate($email : String!, $password: String!) {
       expiresAt
       accessToken
     }
+    customerUserErrors {
+      code
+      field
+      message
+    }
   }
 }
 ''';