+++ lib/mixins/mixins.dart
@@ -1,0 +1,1 @@
+export 'src/shopify_error.dart';