+++ lib/models/src/cart/inputs/cart_inputs.dart
@@ -1,3 +1,4 @@
+export 'attribute_input/attribute_input.dart';
 export 'cart_buyer_identity_input/cart_buyer_identity_input.dart';
 export 'cart_input/cart_input.dart';
 export 'cart_line_update_input/cart_line_update_input.dart';